requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=1.5
tqdm>=4.41.0
seaborn>=0.11.0
pandas
thop
pycocotools
ensemble-boxes
albumentations